+++ src/collection.ts
@@ -39,3 +39,19 @@ export function findFirst<T>(collection: T[], condition: CollectionConditionCall
     }
     return [collection[idx], idx];
 }
+
+/**
+ * Handy when argument can be either single element or an array.
+ * @param t single element or array of the elements
+ * @param copy? Set to true if array argument shall be copied
+ * @returns array of the elements
+ */
+export function makeArray<T>(t?: T | T[], copy?: boolean): T[] {
+    if (!t) {
+        return [];
+    }
+    if (Array.isArray(t)) {
+        return copy === true ? t : [...t];
+    }
+    return [t];
+}